requirements.txt
pandas>=2.2
openpyxl>=3.1
requests>=2.31
python-dotenv>=1.0
